+++ cpp/src/fedpro/session/buffers/UnboundedBuffer.h
@@ -22,6 +22,7 @@
 
 #include <mutex>
 #include <queue>
+#include <cstddef>
 
 namespace FedPro
 {
@@ -33,7 +34,7 @@ namespace FedPro
 
       using peek_type = typename GenericBuffer<E>::peek_type;
 
-      using size_type = uint32_t;
+      using size_type = size_t;
 
       /**
       * @param mutex Used to signal the instance of this class to stop waiting for insertion or poll.
